% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_agency_type}
\alias{check_agency_type}
\title{Check agency type validity}
\usage{
check_agency_type(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check agency type validity
}
\keyword{internal}
