% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_api_url}
\alias{check_api_url}
\title{Check API URL validity}
\usage{
check_api_url(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check API URL validity
}
\keyword{internal}
