% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_api_version}
\alias{check_api_version}
\title{Check API version validity}
\usage{
check_api_version(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check API version validity
}
\keyword{internal}
