% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_award_codes}
\alias{check_award_codes}
\title{Check award codes validity}
\usage{
check_award_codes(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check award codes validity
}
\keyword{internal}
