% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_endpoint}
\alias{check_endpoint}
\title{Check endpoint validity}
\usage{
check_endpoint(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check endpoint validity
}
\keyword{internal}
