% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_fiscal_year}
\alias{check_fiscal_year}
\title{Check fiscal year validity}
\usage{
check_fiscal_year(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check fiscal year validity
}
\keyword{internal}
