% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_fiscal_year_optional}
\alias{check_fiscal_year_optional}
\title{Validate an optional fiscal year}
\usage{
check_fiscal_year_optional(x)
}
\arguments{
\item{x}{Value to validate}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Validate an optional fiscal year
}
