% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_page_all}
\alias{check_page_all}
\title{Check page_all parameter validity}
\usage{
check_page_all(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check page_all parameter validity
}
\keyword{internal}
