% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_page_limit}
\alias{check_page_limit}
\title{Check limit validity}
\usage{
check_page_limit(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check limit validity
}
\keyword{internal}
