% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_page_number}
\alias{check_page_number}
\title{Check page number validity}
\usage{
check_page_number(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check page number validity
}
\keyword{internal}
