% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_response_type}
\alias{check_response_type}
\title{Check response content type validity}
\usage{
check_response_type(x)
}
\arguments{
\item{x}{Response to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check response content type validity
}
\keyword{internal}
