% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check_toptier_code}
\alias{check_toptier_code}
\title{Check toptier code validity}
\usage{
check_toptier_code(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Invisibly returns TRUE if valid, errors if not
}
\description{
Check toptier code validity
}
\keyword{internal}
