% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{has_pagination}
\alias{has_pagination}
\title{Check if a response has pagination}
\usage{
has_pagination(x)
}
\arguments{
\item{x}{Response to check}
}
\value{
TRUE if response has pagination metadata
}
\description{
Check if a response has pagination
}
\keyword{internal}
