% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{has_valid_award_codes}
\alias{has_valid_award_codes}
\title{Check if value has valid award type codes}
\usage{
has_valid_award_codes(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Logical indicating if x contains valid award type codes
}
\description{
Check if value has valid award type codes
}
\keyword{internal}
