% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{has_valid_content_type}
\alias{has_valid_content_type}
\title{Check if response has valid content type}
\usage{
has_valid_content_type(x)
}
\arguments{
\item{x}{Response to check}
}
\value{
TRUE if response has valid content type
}
\description{
Check if response has valid content type
}
\keyword{internal}
