% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_agency_type}
\alias{is_agency_type}
\title{Check if a string is a valid agency type}
\usage{
is_agency_type(x)
}
\arguments{
\item{x}{String to test}
}
\value{
TRUE if valid agency type
}
\description{
Check if a string is a valid agency type
}
