% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_api_url}
\alias{is_api_url}
\title{Check if a string is a valid API URL}
\usage{
is_api_url(x)
}
\arguments{
\item{x}{String to test}
}
\value{
TRUE if valid API URL
}
\description{
Check if a string is a valid API URL
}
\keyword{internal}
