% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_api_version}
\alias{is_api_version}
\title{Check if a string is a valid API version}
\usage{
is_api_version(x)
}
\arguments{
\item{x}{String to test}
}
\value{
TRUE if valid API version
}
\description{
Check if a string is a valid API version
}
\keyword{internal}
