% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_endpoint}
\alias{is_endpoint}
\title{Check if a string is a valid endpoint}
\usage{
is_endpoint(x)
}
\arguments{
\item{x}{String to test}
}
\value{
TRUE if valid endpoint
}
\description{
Check if a string is a valid endpoint
}
\keyword{internal}
