% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_fiscal_year}
\alias{is_fiscal_year}
\title{Check if a value is a valid fiscal year}
\usage{
is_fiscal_year(x)
}
\arguments{
\item{x}{Value to test}
}
\value{
TRUE if valid fiscal year
}
\description{
Check if a value is a valid fiscal year
}
