% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_fiscal_year_optional}
\alias{is_fiscal_year_optional}
\title{Validate an optional fiscal year}
\usage{
is_fiscal_year_optional(x)
}
\arguments{
\item{x}{Value to test}
}
\value{
TRUE if NULL or valid fiscal year
}
\description{
Validate an optional fiscal year
}
