% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_page_all}
\alias{is_page_all}
\title{Check if value is a valid page_all parameter}
\usage{
is_page_all(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
TRUE if valid page_all parameter
}
\description{
Check if value is a valid page_all parameter
}
\keyword{internal}
