% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_page_limit}
\alias{is_page_limit}
\title{Check if value is a valid limit}
\usage{
is_page_limit(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Logical indicating if x is a valid limit
}
\description{
Check if value is a valid limit
}
\keyword{internal}
