% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_page_number}
\alias{is_page_number}
\title{Check if value is a valid page number}
\usage{
is_page_number(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Logical indicating if x is a valid page number
}
\description{
Check if value is a valid page number
}
\keyword{internal}
