% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_page_range}
\alias{is_page_range}
\title{Check if value is a valid page range}
\usage{
is_page_range(x, total_pages)
}
\arguments{
\item{x}{Value to check}

\item{total_pages}{Maximum number of pages available}
}
\value{
Logical indicating if x is a valid page range
}
\description{
Check if value is a valid page range
}
\keyword{internal}
