% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_sort_direction}
\alias{is_sort_direction}
\title{Check if value is a valid sort direction}
\usage{
is_sort_direction(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Logical indicating if x is a valid sort direction
}
\description{
Check if value is a valid sort direction
}
\keyword{internal}
