% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{is_toptier_code}
\alias{is_toptier_code}
\title{Check if a string is a valid toptier code}
\usage{
is_toptier_code(x)
}
\arguments{
\item{x}{String to test}
}
\value{
TRUE if valid toptier code
}
\description{
Check if a string is a valid toptier code
}
