% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{usasp}
\alias{usasp}
\title{Make a call to the USASpending API}
\usage{
usasp(
  endpoint,
  ...,
  .api_url = "https://api.usaspending.gov",
  .api_version = "v2",
  .page_all = FALSE,
  .limit = 10,
  .as_tibble = TRUE,
  .flatten = FALSE
)
}
\arguments{
\item{endpoint}{API endpoint to call}

\item{...}{Name-value pairs giving API parameters}

\item{.api_url}{The base URL for the API}

\item{.api_version}{The API version to use}

\item{.page_all}{Logical, whether to automatically fetch all pages}

\item{.limit}{Number of results per page (max 100)}

\item{.as_tibble}{Logical, whether to return response as a tibble}

\item{.flatten}{Logical, whether to flatten nested structures in tibble output}
}
\value{
Response from the API, either as a tibble or raw response
}
\description{
Make a call to the USASpending API
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Get single page of results as a tibble
usasp("/agency/awards/count", fiscal_year = 2023)

# Get all pages of flattened results
usasp("/agency/awards/count", fiscal_year = 2023, 
      .page_all = TRUE, .flatten = TRUE)
      
# Get raw JSON response
usasp("/agency/awards/count", fiscal_year = 2023, .as_tibble = FALSE)
\dontshow{\}) # examplesIf}
}
