% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usaspending-package.R
\docType{package}
\name{usaspending-package}
\alias{usaspending}
\alias{usaspending-package}
\title{Interface to USAspending.gov API}
\description{
Provides a comprehensive interface to access federal spending data through the
USAspending.gov API (v2). The package implements endpoints for retrieving data
about agencies, awards, recipients, and reference data, with consistent return
formats and error handling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-pkg.thecoatlessprofessor.com/usaspending/}
  \item \url{https://github.com/coatless-rpkg/usaspending}
  \item Report bugs at \url{https://github.com/coatless-rpkg/usaspending/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com}

}
