% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-api.R
\name{print.usasp_response}
\alias{print.usasp_response}
\title{Print Method for Marked Up USAspending API Responses}
\usage{
\method{print}{usasp_response}(x, ...)
}
\arguments{
\item{x}{A raw API response of class \code{'usasp_response'}}

\item{...}{Additional arguments passed to print}
}
\description{
Provides a structured display of API response data
}
